requirements.txt
appnope==0.1.2
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
blurhash==1.1.4
boost==0.1
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
cycler==0.11.0
decorator==5.1.1
deepdish==0.3.7
dlib==19.23.0
executing==0.8.2
Flask==2.0.2
Flask-Cors==3.0.10
fonttools==4.29.0
greenlet==1.1.2
h5py==3.6.0
idna==3.3
imageio==2.14.1
ipython==8.0.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
Mastodon.py==1.5.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
networkx==2.6.3
nibabel==3.2.1
nilearn==0.9.0
nltools==0.4.5
numexpr==2.8.1
numpy==1.22.1
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
openface==0.2.1
packaging==21.3
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
prompt-toolkit==3.0.26
ptyprocess==0.7.0
pure-eval==0.2.2
py-feat==0.3.7
Pygments==2.11.2
pynv==0.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-magic==0.4.24
pytz==2021.3
PyWavelets==1.2.0
requests==2.27.1
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.31
stack-data==0.1.4
tables==3.7.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tomli==2.0.0
torch==1.10.2
torchvision==0.11.3
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.2